backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
pydantic-settings>=2.1.0
email-validator>=2.0.0
alembic>=1.13.1
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
pgvector>=0.2.4
httpx>=0.26.0
# AI/ML Dependencies
numpy>=1.26.0
scikit-learn>=1.4.0
torch>=2.2.0
torchvision>=0.17.0
facenet-pytorch>=2.5.3
pillow>=10.2.0
opencv-python-headless>=4.9.0
